apps/backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
bcrypt==5.0.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dotenv==0.9.9
executing==2.2.1
fastapi==0.129.2
filetype==1.2.0
fonttools==4.61.1
h11==0.16.0
idna==3.7
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jose==1.0.0
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.2
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
packaging==26.0
parso==0.8.6
passlib==1.7.4
pexpect==4.9.0
pi_heif==1.3.0
pillow==12.1.1
pillow-avif-plugin==1.5.5
platformdirs==4.9.2
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
pyDeprecate==0.5.0
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.15
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.52.1
supervision==0.27.0.post1
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
